% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{initialize_NH}
\alias{initialize_NH}
\title{Initialize nursing home}
\usage{
initialize_NH(
  p_asymp_nonres = 0.5,
  p_asymp_res = 0.5,
  p_subclin_nonres = 0,
  p_subclin_res = 0,
  daily_attack = 0.18,
  staff_vax_req = F,
  res_boost = 0.74,
  staff_boost = 0.51,
  visit_boost = 0.34,
  staff_trans_red = 0.7,
  visit_trans_red = 0,
  res_trans_red = 0,
  staff_susp_red = 0,
  visit_susp_red = 0,
  res_susp_red = 0,
  isolate = T,
  prim_previnf_eff = 0.4,
  boost_eff = 0.7,
  days_inf = 5,
  days_isolate_res = 10,
  days_isolate_staff = 7,
  days_isolate_visit = 5,
  start
)
}
\arguments{
\item{p_asymp_nonres}{Fraction of non-residents with asymptomatic disease; defaults to 0.5}

\item{p_asymp_res}{Fraction of residents with asymptomatic disease; defaults to 0.5}

\item{p_subclin_nonres}{Fraction of non-residents with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{p_subclin_res}{Fraction of residents with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{daily_attack}{Average daily attack rate; defaults to 0.18}

\item{staff_vax_req}{Whether staff are required to get most updated recommended vaccine; defaults to F}

\item{res_boost}{Proportion of residents with booster dose}

\item{staff_boost}{Proportion of staff with booster dose}

\item{visit_boost}{Proportion of visitors with booster dose}

\item{staff_trans_red}{Factor by which staff transmissibility is reduced due to intervention; defaults to 0.7
(used to be teacher_trans)}

\item{visit_trans_red}{Factor by which visitor transmissibility is reduced due to intervention; defaults to 0
(used to be family_susp)}

\item{res_trans_red}{Factor by which resident transmissibility is reduced due to intervention; defaults to 0}

\item{staff_susp_red}{Factor by which staff susceptibility is reduced due to intervention; defaults to 0
(used to be teacher_trans)}

\item{visit_susp_red}{Factor by which visitor susceptibility is reduced due to intervention; defaults to 0
(used to be family_susp)}

\item{res_susp_red}{Factor by which resident susceptibility is reduced due to intervention; defaults to 0}

\item{isolate}{Whether symptomatic individuals isolate when symptoms emerge; defaults to T}

\item{prim_previnf_eff}{efficacy of 2 doses or previous infection; defaults to 0.4}

\item{boost_eff}{booster efficacy; defaults to 0.7}

\item{days_isolate_res}{length of isolation for residents, defaults to 10}

\item{days_isolate_staff}{length of isolation for staff, defaults to 7}

\item{days_isolate_visit}{length of isolation for visitors, defaults to 5}

\item{start}{dataframe from make_NH()}
}
\value{
df dataframe of resident and staff attributes
}
\description{
This function takes in a data frame exported by make_NH().
It adds epidemiological attributes of the full nursing home community.
}
