% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{make_room}
\alias{make_room}
\title{Create list of people in each room}
\usage{
make_room(df)
}
\arguments{
\item{df}{data frame from make_schedule()}
}
\value{
cohorts list of people in each staff/resident/visitor room
}
\description{
Store info about people in each room per shift per day
}
