% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{make_schedule}
\alias{make_schedule}
\title{Make schedule}
\usage{
make_schedule(time = 45, nh)
}
\arguments{
\item{time}{number of days; defaults to 45 (to capture 30-day picture)}

\item{nh}{dataframe from initialize_NH()}
}
\value{
d Returns an n x time data frame that indicates whether an individual is in the 
nursing home at a particular time
}
\description{
Make a schedule of when staff and visitors are present/absent
}
