% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{mult_runs}
\alias{mult_runs}
\title{Run model multiple times and summarize results}
\usage{
mult_runs(
  N,
  rel_trans_common = 1/4,
  rel_trans_staff = 1/4,
  p_asymp_nonres = 0.5,
  p_asymp_res = 0.5,
  p_subclin_nonres = 0,
  p_subclin_res = 0,
  daily_attack = 0.18,
  staff_vax_req = F,
  res_boost = 0,
  staff_boost = 0,
  visit_boost = 0,
  staff_trans_red = 0.7,
  visit_trans_red = 0,
  res_trans_red = 0,
  staff_susp_red = 0,
  visit_susp_red = 0,
  res_susp_red = 0,
  n_contact_common_res = 3,
  n_contact_common_staff = 3,
  n_contact_staff = 6,
  days_inf = 5,
  days_isolate_res = 10,
  days_isolate_staff = 7,
  days_isolate_visit = 5,
  seed_asymp = F,
  isolate = T,
  time = 30,
  test = F,
  test_sens = 0.84,
  test_frac = 0.9,
  test_days = "week",
  test_type = "all",
  test_start_day = 7,
  visit_test = F,
  staff_prob = 0.005,
  visit_prob = 0.005,
  quarantine = F,
  quarantine.length = 7,
  prim_previnf_eff = 0.4,
  boost_eff = 0.7,
  test_recovered = T,
  sched,
  cohorts
)
}
\arguments{
\item{N}{number of runs}

\item{rel_trans_common}{Relative attack rate of common area contact (vs. room); defaults to 1/4}

\item{rel_trans_staff}{Relative attack rate of staff-staff interactions (vs. room); defaults to 1/4}

\item{p_asymp_nonres}{Fraction of staff with asymptomatic (unsuspected) disease; defaults to 0.5}

\item{p_asymp_res}{Fraction of residents with asymptomatic (unsuspected) disease; defaults to 0.5}

\item{p_subclin_nonres}{Fraction of non-residents with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{p_subclin_res}{Fraction of residents with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{daily_attack}{Average daily attack rate in residents; defaults to 0.18}

\item{staff_vax_req}{Whether staff are required to get most updated vaccine; defaults to F}

\item{res_boost}{Proportion of residents with booster dose}

\item{staff_boost}{Proportion of staff with booster dose}

\item{visit_boost}{Proportion of visitors with booster dose}

\item{staff_trans_red}{Factor by which staff transmission is reduced due to intervention; defaults to 0.7}

\item{visit_trans_red}{Factor by which visitor transmission is reduced due to intervention; defaults to 0}

\item{res_trans_red}{Factor by which resident transmission is reduced due to intervention; defaults to 0}

\item{staff_susp_red}{Factor by which staff susceptibility is reduced due to intervention; defaults to 0}

\item{visit_susp_red}{Factor by which visitor susceptibility is reduced due to intervention; defaults to 0}

\item{res_susp_red}{Factor by which resident susceptibility is reduced due to intervention; defaults to 0}

\item{n_contact_common_res}{number of contacts a staff/resident has with other residents in common area; defaults to 3}

\item{n_contact_common_staff}{number of contacts a staff/resident has with other staff in common area; defaults to 3}

\item{n_contact_staff}{number of contacts a staff member has with other staff members; defaults to 6}

\item{days_inf}{length of infectious period, defaults to 5}

\item{days_isolate_res}{length of isolation for residents, defaults to 10}

\item{days_isolate_staff}{length of isolation for staff, defaults to 7}

\item{days_isolate_visit}{length of isolation for visitors, defaults to 5}

\item{seed_asymp}{whether to seed with an asymptomatic case}

\item{isolate}{Whether symptomatic individuals isolate when symptoms emerge; defaults to T}

\item{time}{length of time to run model; defaults to 30}

\item{test}{whether there is weekly testing; defaults to F}

\item{test_sens}{test sensitivity; defaults to 0.84}

\item{test_frac}{fraction of school tested; defaults to 0.9}

\item{test_days}{test frequency; "day", "week", "2x_week"}

\item{test_type}{group tested; defaults to "all", also allows "staff" and "residents"}

\item{test_start_day}{day tests are implemented for weekly testing}

\item{visit_test}{whether visitors test before entering NH}

\item{staff_prob}{set daily probability of infectious entry for staff, based on community incidence}

\item{visit_prob}{set daily probability of infectious entry for visitors, based on community incidence}

\item{quarantine}{whether people quarantine upon exposure; defaults to F}

\item{quarantine.length}{length of quarantine when someone is infectious; defaults to 7}

\item{prim_previnf_eff}{efficacy of 2 doses or previous infection; defaults to 0.4}

\item{boost_eff}{booster efficacy; defaults to 0.7}

\item{test_recovered}{whether or not to test recovered individuals; defaults to T}

\item{sched}{nursing home population schedule; from make_schedule()}

\item{cohorts}{list from make_room()}
}
\description{
Run model multiple times and summarize results
}
