% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{run_common}
\alias{run_common}
\title{Set common area transmission (used to be run_rand)}
\usage{
run_common(
  a,
  df,
  n_contact_common_res,
  n_contact_common_staff,
  rel_trans_common = 1/4
)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{data frame in run_model()}

\item{n_contact_common_res}{number of contacts a staff/residents has with other residents in the common area; defaults to 3}

\item{n_contact_common_staff}{number of contacts a staff/residents has with other staff in the common area; defaults to 6}

\item{rel_trans_staff}{relative transmission in common area interactions (vs. resident room);
defaults to 1/4}
}
\value{
id's of infected individuals
}
\description{
Determine who is infected at a timestep
from common area contact with an infected individual
}
