% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{run_model}
\alias{run_model}
\title{Run model}
\usage{
run_model(
  time = 30,
  test = F,
  test_days = "week",
  test_sens = 0.84,
  test_frac = 0.9,
  test_start_day = 7,
  visit_test = F,
  n_contact_common_res = 3,
  n_contact_common_staff = 3,
  n_contact_staff = 6,
  days_inf = 5,
  days_isolate_res = 10,
  days_isolate_staff = 7,
  days_isolate_visit = 5,
  seed_asymp = F,
  staff_prob = 0.005,
  visit_prob = 0.005,
  quarantine = F,
  quarantine.length = 7,
  rel_trans_common = 1/4,
  rel_trans_staff = 1/4,
  test_type = "all",
  test_recovered = T,
  df,
  cohorts
)
}
\arguments{
\item{time}{length of time to run model; defaults to 30}

\item{test}{whether there is screening}

\item{test_days}{test frequency; "day", "week", "2x_week"; defaults to "week"}

\item{test_sens}{test sensitivity; defaults to 0.84}

\item{test_frac}{fraction of nursing home tested; defaults to 0.9}

\item{test_start_day}{day tests are implemented for weekly testing}

\item{visit_test}{whether visitors test before entering NH}

\item{n_contact_common_res}{number of contacts a staff/resident has with other residents in the common area; defaults to 3}

\item{n_contact_common_staff}{number of contacts a staff/resident has with other staff in the common area; defaults to 3}

\item{n_contact_staff}{number of contacts a staff member has with other staff members during a shift; defaults to 6}

\item{days_inf}{length of infectious period, defaults to 5 days}

\item{days_isolate_res}{length of isolation for residents, defaults to 10}

\item{days_isolate_staff}{length of isolation for staff, defaults to 7}

\item{days_isolate_visit}{length of isolation for visitors, defaults to 5}

\item{seed_asymp}{whether to seed with an asymptomatic case}

\item{staff_prob}{set daily probability of infectious entry for staff (based on COVID-19 community incidence)}

\item{visit_prob}{set daily probability of infectious entry for visitors (based on COVID-19 community incidence)}

\item{quarantine}{whether or not people quarantine upon exposure; defaults to F}

\item{quarantine.length}{length of quarantine when someone is exposed; defaults to 7}

\item{rel_trans_common}{relative transmission in common area interactions vs. resident's room; defaults to 1/4}

\item{rel_trans_staff}{relative transmission in staff-staff interactions vs. resident's room; defaults to 1/4}

\item{test_type}{group tested; defaults to "all", also allows "residents" and "staff"}

\item{test_recovered}{whether or not to test recovered individuals; defaults to T}

\item{df}{data frame from make_schedule()}

\item{cohorts}{list from make_room()}
}
\value{
df updated df with transmission results

time_seed_inf when the first individual was dropped in
}
\description{
Perform a single model run
}
