% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{run_room}
\alias{run_room}
\title{Set room transmission (used to be run_household)}
\usage{
run_room(a, df, t, quarantine, cohorts, shift)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{data frame in run_model()}

\item{t}{current day}

\item{quarantine}{whether quarantine occurs}

\item{cohorts}{list of dataframes of people in the same room as each person}

\item{shift}{current shift}
}
\value{
id's of infected individuals
}
\description{
Determine who is infected at a timestep
in the same room as an infected individual
}
