% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_abm_master.R
\name{run_staff}
\alias{run_staff}
\title{Set staff transmission}
\usage{
run_staff(a, df, n_contact_staff, rel_trans_staff = 1/4)
}
\arguments{
\item{a}{id of infected staff member}

\item{df}{school data frame in run_model()}

\item{n_contact_staff}{number of contacts staff encounters in nursing home during shift}

\item{rel_trans_staff}{relative transmission in staff-staff interactions (vs. resident room);
defaults to 1/4}
}
\value{
id's of infected individuals
}
\description{
Determine who is infected at a timestep from contact between in-nursing-home staff
}
